Refactoring
Busted bigger class into something smaller

Issue #384
